library/pyproject.toml
[project]
name = "library"
version = "0.0.1"
description = "An application for the library data ingestion"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.0",
    "botocore>=1.41.6",
    "dependency-injector>=4.46.0",
    "icecream>=2.1.4",
    "logfire>=3.14.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.11",
    "pyalex>=0.19",
    "pyarrow>=22.0.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.3",
    "sqlmodel>=0.0.27",
    "tqdm>=4.67.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src/library"]


[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.8",
]

gpu = [
    "setfit>=1.1.3",
    "scikit-learn==1.6.1",
    "ollama>=0.4.7",
]

pdfscraping = [
    "pymupdf4llm>=0.0.17",
    "pymupdf-layout>=1.26.6",
    "opencv-python-headless>=4.12.0.88",
]

webscraping = [
    "selenium>=4.29.0",
    "webdriver-manager>=4.0.2",
]
